# [5.8] Arr::flatten performance