MacBook Pro (2017), i5, 16 GB

**Before**: 1.92s
**After**: 6.5ms
```php
$arr = [];
for ($i = 0; $i < 20000; $i++) {
    $arr[] = ['test'];
}
Arr::flatten($arr);
```

**Before**: 6.33s
**After**: 87ms
```php
$arr = [];
for ($i = 0; $i < 20000; $i++) {
    $arr[] = ['name' => 'John', 'email' => 'john@me.com'];
}
app('db')->table('users')->insert($arr);
```

**Before**: 5.7s
**After**: 88.93ms
```php
$arr = [];
for ($i = 0; $i < 20000; $i++) {
    $arr[] = ['name' => 'John', 'email' => 'john@me.com'];;
}
User::insert($arr);
```