@bbsnly 2019-05-26
@laurencei I was comparing the `insert` performance between Eloquent and Doctrine, to do so I prepared ten different ways to insert new data with Laravel (Eloquent, QueryBuilder, direct connection). At some point I saw that when doing bulk insertion with thousands of items the performance was very poor.
When debugging I found the bottleneck here `Illuminate/Database/Query/Builder::insert` - `Arr::flatten`. It worked perfectly with a low amount of items but when using it with tens of thousands of items it was very slow.
The explanation of my solution is very easy: `array_merge` is slow when merging to a big array as it looks and compares keys where the concatenation operation is very fast. We are not interested in keys (`array_values` is used) so by using the `array_merge` function we will perform many operations we do not need.