@mfn 2019-05-27
Very nice improvements, thanks @bbsnly 

> `array_merge` is slow when merging to a big array

Can confirm, in my experience the 2nd worst bottleneck (1st being the database 😉) I've seen in practice (and: being easily identifiable) is this function. If used in a loop and large datasets, almost always a recipe to avoid.