@mfn 2019-05-29
> array_push($result, ...$values)

wow, that's neat trick!

I wonder if there are downsides or limitations of argument unpacking a large set for a function call?