@solofeed 2019-05-29
@mfn We can add a check for the size of the array and decide what to use

@bbashy What you think? maybe you can improve speed more?)