@ad-mos 2019-05-31
Looks like I am doing something wrong)
Tested by merging 2 arrays with 5M elements each.
```
foreach ... $old[] = $item;
0.58910183906555

array_push($old, ...$new);
0.52405204772949

array_merge($old, $new);
0.36471142768860
```

Here is the code:
```
$count = 10000000;
$iter = 5;

$resInsert = 0.0;
$resPush = 0.0;
$resMerge = 0.0;

for ($i = 0; $i < $iter; $i++) {
    $resInsert += testInsert($count);
    $resPush += testPush($count);
    $resMerge += testMerge($count);
}

echo 'foreach ... $old[] = $item;' . "\n";
echo sprintf("%.14f\n\n", $resInsert / $iter);

echo 'array_push($old, ...$new);' . "\n";
echo sprintf("%.14f\n\n", $resPush / $iter);

echo 'array_merge($old, $new);' . "\n";
echo sprintf("%.14f\n\n", $resMerge / $iter);

die;

function testPush($count)
{
    [$old, $new] = getArrays($count);

    $t1 = microtime(true);
    array_push($old, ...$new);
    $t2 = microtime(true);

    unset($old);
    unset($new);

    return $t2 - $t1;
}

function testMerge($count)
{
    [$old, $new] = getArrays($count);

    $t1 = microtime(true);
    $old = array_merge($old, $new);
    $t2 = microtime(true);

    unset($old);
    unset($new);

    return $t2 - $t1;
}

function testInsert($count)
{
    [$old, $new] = getArrays($count);

    $t1 = microtime(true);
    foreach ($new as $item) {
        $old[] = $item;
    }
    $t2 = microtime(true);

    unset($old);
    unset($new);

    return $t2 - $t1;
}

function getArrays(int $count)
{
    $halfCount = floor($count / 2);

    $old = [];
    for ($i = $halfCount; $i > 0; $i--) {
        $old[\Illuminate\Support\Str::random(3) . $i] = 'a';
    }

    $new = array_fill(0, $halfCount, 'a');

    return [$old, $new];
}
```